Добавлены классы для джобов

+++ MetricsManager/DAL/Migrations/DBMigration.cs
@@ -12,7 +12,7 @@ namespace MetricsManager.DAL.Migrations
         public override void Up()
         {
             Create.Table("agents")
-                .WithColumn("AgentId").AsInt64().PrimaryKey().Identity()
+                .WithColumn("AgentId").AsInt32().PrimaryKey().Identity()
                 .WithColumn("AgentAddress").AsString();
 
             Create.Table("cpumetrics")
